moved dependencies to pkg folder
