+++ pkg/certutil/certificate.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	logger "github.com/jsandas/gologger"
-	"github.com/jsandas/tlstools/utils"
+	"github.com/jsandas/tlstools/pkg/utils"
 )
 
 // CertData certificate data fields
