+++ pkg/certutil/common.go
