+++ pkg/ssl/common.go
