+++ pkg/vuln/weakkey/weakkey.go
