+++ pkg/vuln/weakkey/weakkey_test.go
